refactor: move URI resolver to new package
