+++ mycore-base/src/main/java/org/mycore/common/xml/MCRFunctionResolver.java
@@ -30,6 +30,7 @@ import javax.xml.transform.URIResolver;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.mycore.common.MCRClassTools;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 
 /**
  * Resolves arbitrary static methods of arbitrary classes. Parameters are considerd to be of type
