+++ mycore-base/src/test/java/org/mycore/common/xml/MCRRateLimitResolverTest.java
@@ -37,6 +37,7 @@ import org.mycore.common.MCRException;
 import org.mycore.common.MCRRateLimitBuckets;
 import org.mycore.common.config.MCRConfiguration2;
 import org.mycore.common.config.MCRConfigurationException;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 import org.mycore.test.MyCoReTest;
 
 import io.github.bucket4j.Bucket;
