+++ mycore-mods/src/main/java/org/mycore/mods/enrichment/MCREnrichmentResolver.java
@@ -23,7 +23,7 @@ import javax.xml.transform.URIResolver;
 
 import org.jdom2.Element;
 import org.jdom2.transform.JDOMSource;
-import org.mycore.common.xml.MCRURIResolver;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 
 /**
  * Retrieves a publication as MODS XML from a given URI and
