+++ mycore-mods/src/test/java/org/mycore/mods/enrichment/MCREnrichmentTest.java
@@ -29,8 +29,8 @@ import org.jdom2.output.XMLOutputter;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.mycore.common.xml.MCRNodeBuilder;
-import org.mycore.common.xml.MCRURIResolver;
 import org.mycore.common.xml.MCRXMLHelper;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 import org.mycore.test.MyCoReTest;
 
 @MyCoReTest
