+++ mycore-user2/src/test/java/org/mycore/user2/MCRUserManagerTest.java
@@ -38,7 +38,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mycore.common.MCRTestConfiguration;
 import org.mycore.common.MCRTestProperty;
-import org.mycore.common.xml.MCRURIResolver;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 import org.mycore.test.MCRJPAExtension;
 import org.mycore.test.MCRJPATestHelper;
 import org.mycore.test.MyCoReTest;
