+++ mycore-wfc/src/main/java/org/mycore/wfc/actionmapping/MCRActionMappingURIResolver.java
@@ -26,7 +26,7 @@ import javax.xml.transform.Source;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.URIResolver;
 
-import org.mycore.common.xml.MCRURIResolver;
+import org.mycore.common.xsl.uriresolver.MCRURIResolver;
 import org.mycore.services.http.MCRURLQueryParameter;
 
 /**
